Use Github hosted runners for Actions
As we moved to VPC based Lambdas, but have kept the isolation, we now
have cold build times exceeding 30 minutes, that is tolerable in itself,
but the implication of only having a single runner running a single step
at a time is creating a large amount of congestion in the CI.
Using Github Hosted runners, authenticating by an Service Account IAM
User, allow for concurrency in the terraform steps, relieving a large
amount of the congestion.
Also after doing some analysis and comparison, GIthub Actions is a much
more mature and functional CI platform, compared to AWS Code offerings.
Therefore in the interest of a low management cost CI tool, I think
Github Actions should be the prefered choice. This is with the current
constraint of deployments to production being run by a human
administrator, not by CI.