Adding timeouts to PDS API requests
Without this, we see the lambdas hanging forever, up until their
configured timeouts, under error conditions. Which mkaes debugging
difficult, especially as log messages may not be returned when the
Lambda is terminated by the run time, leading engineers to be
guessing as to what caused the request to not complete.