PCRM-4392 Lambdas in shared Account VPC
This will reduce build times and costs, at the trade off of not having
completely isolated environments in the dev account.
We were seeing build times of over 40 minutes when terraform had to
provision a new VPC and supporting networking infrastructure per branch.
This is mostly caused by the Lambda provisioning time having to create
an ENI per lambda, which this does not change and there doesn't appear
to be anything we can do to improve it, hopefully AWS can make some
platform improvements in this area.
The builds have improved by ~10 minutes, to around 30 minutes from not
having to provision the VPC and it's supporting infrastructure.